+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/DefaultMediaSourceFactory.java
@@ -238,16 +238,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
     return this;
   }
 
-  @CanIgnoreReturnValue
-  @Override
-  @UnstableApi
-  public DefaultMediaSourceFactory experimentalSetCodecsToParseWithinGopSampleDependencies(
-      @C.VideoCodecFlags int codecsToParseWithinGopSampleDependencies) {
-    delegateFactoryLoader.setCodecsToParseWithinGopSampleDependencies(
-        codecsToParseWithinGopSampleDependencies);
-    return this;
-  }
-
   /**
    * Sets the {@link AdsLoader.Provider} that provides {@link AdsLoader} instances for media items
    * that have {@link MediaItem.LocalConfiguration#adsConfiguration ads configurations}.
@@ -721,7 +711,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
     private DataSource.@MonotonicNonNull Factory dataSourceFactory;
     private boolean parseSubtitlesDuringExtraction;
     private SubtitleParser.Factory subtitleParserFactory;
-    private @C.VideoCodecFlags int codecsToParseWithinGopSampleDependencies;
     private boolean loadOnlySelectedTracks;
     @Nullable private CmcdConfiguration.Factory cmcdConfigurationFactory;
     @Nullable private DrmSessionManagerProvider drmSessionManagerProvider;
@@ -736,7 +725,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
       mediaSourceFactories = new HashMap<>();
       parseSubtitlesDuringExtraction = true;
       loadOnlySelectedTracks = true;
-      codecsToParseWithinGopSampleDependencies = C.VIDEO_CODEC_FLAG_H264 | C.VIDEO_CODEC_FLAG_H265;
     }
 
     public @C.ContentType int[] getSupportedTypes() {
@@ -768,8 +756,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
       }
       mediaSourceFactory.setSubtitleParserFactory(subtitleParserFactory);
       mediaSourceFactory.experimentalParseSubtitlesDuringExtraction(parseSubtitlesDuringExtraction);
-      mediaSourceFactory.experimentalSetCodecsToParseWithinGopSampleDependencies(
-          codecsToParseWithinGopSampleDependencies);
       mediaSourceFactories.put(contentType, mediaSourceFactory);
       return mediaSourceFactory;
     }
@@ -799,13 +785,6 @@ public final class DefaultMediaSourceFactory implements MediaSourceFactory {
       for (MediaSource.Factory mediaSourceFactory : mediaSourceFactories.values()) {
         mediaSourceFactory.setSubtitleParserFactory(subtitleParserFactory);
       }
-    }
-
-    public void setCodecsToParseWithinGopSampleDependencies(
-        @C.VideoCodecFlags int codecsToParseWithinGopSampleDependencies) {
-      this.codecsToParseWithinGopSampleDependencies = codecsToParseWithinGopSampleDependencies;
-      extractorsFactory.experimentalSetCodecsToParseWithinGopSampleDependencies(
-          codecsToParseWithinGopSampleDependencies);
     }
 
     public void setCmcdConfigurationFactory(CmcdConfiguration.Factory cmcdConfigurationFactory) {