+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/MediaSource.java
@@ -33,7 +33,6 @@ import androidx.media3.exoplayer.upstream.BandwidthMeter;
 import androidx.media3.exoplayer.upstream.CmcdConfiguration;
 import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy;
 import androidx.media3.exoplayer.util.ReleasableExecutor;
-import androidx.media3.extractor.mp4.Mp4Extractor;
 import androidx.media3.extractor.text.SubtitleParser;
 import com.google.common.base.Supplier;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
@@ -135,27 +134,6 @@ public interface MediaSource {
      */
     @UnstableApi
     default Factory setSubtitleParserFactory(SubtitleParser.Factory subtitleParserFactory) {
-      return this;
-    }
-
-    /**
-     * Sets the set of video codecs for which within GOP sample dependency information should be
-     * parsed as part of extraction. Defaults to H.264 and H.265.
-     *
-     * <p>Having access to additional sample dependency information can speed up seeking. See {@link
-     * Mp4Extractor#FLAG_READ_WITHIN_GOP_SAMPLE_DEPENDENCIES}.
-     *
-     * <p>This method is experimental and will be renamed or removed in a future release.
-     *
-     * @param codecsToParseWithinGopSampleDependencies The set of codecs for which to parse within
-     *     GOP sample dependency information.
-     * @return This factory, for convenience.
-     */
-    @UnstableApi
-    @CanIgnoreReturnValue
-    @ExperimentalApi // TODO: b/470365670 - Remove method once config is enabled by default.
-    default Factory experimentalSetCodecsToParseWithinGopSampleDependencies(
-        @C.VideoCodecFlags int codecsToParseWithinGopSampleDependencies) {
       return this;
     }