+++ RELEASENOTES.md
@@ -62,6 +62,12 @@
         workflows.
 *   Track Selection:
 *   Extractors:
+    *   Remove `Mp4Extractor.FLAG_READ_WITHIN_GOP_SAMPLE_DEPENDENCIES`,
+        `Mp4Extractor.FLAG_READ_WITHIN_GOP_SAMPLE_DEPENDENCIES_H265`,
+        `FragmentedMp4Extractor.FLAG_READ_WITHIN_GOP_SAMPLE_DEPENDENCIES`, and
+        `FragmentedMp4Extractor.FLAG_READ_WITHIN_GOP_SAMPLE_DEPENDENCIES_H265`
+        as within-GOP sample dependency parsing for H.264 and H.265 is enabled
+        by default.
     *   Add `ExtractorUtil.getFramesPerEncodedSample` to calculate the number of
         audio frames per sample for encoded audio formats
         ([#3367](https://github.com/androidx/media/issues/3367)).